# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a server for Slyther Go Backend.",
        "title": "Slyther Go Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/contracts/{address}": {
            "get": {
                "description": "Get a cached smart contract from the database by its address and network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Get contract from database",
                "parameters": [
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "description": "Contract Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "ethereum",
                            "polygon",
                            "bsc",
                            "base",
                            "arbitrum",
                            "avalanche",
                            "optimism",
                            "gnosis",
                            "fantom",
                            "celo"
                        ],
                        "type": "string",
                        "default": "ethereum",
                        "description": "Network Name",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/contracts/{address}/source-code": {
            "get": {
                "description": "Get the source code of a smart contract from Etherscan by its address and save it to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Fetch contract source code",
                "parameters": [
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "description": "Contract Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "ethereum",
                            "polygon",
                            "bsc",
                            "base",
                            "arbitrum",
                            "avalanche",
                            "optimism",
                            "gnosis",
                            "fantom",
                            "celo"
                        ],
                        "type": "string",
                        "default": "ethereum",
                        "description": "Network Name",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/dynamic-analysis": {
            "post": {
                "description": "Performs dynamic security analysis on Solidity source code using AI/LLM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dynamic-analysis"
                ],
                "summary": "Analyze Solidity contract using AI for security vulnerabilities",
                "parameters": [
                    {
                        "description": "Contract source code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/analysis.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/dynamic-analysis/{sourceHash}": {
            "get": {
                "description": "Retrieves dynamic analysis result from cache or database using source hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dynamic-analysis"
                ],
                "summary": "Get dynamic analysis result by source hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source hash",
                        "name": "sourceHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/static-analysis": {
            "post": {
                "description": "Performs static security analysis on Solidity source code using Slither",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "static-analysis"
                ],
                "summary": "Analyze Solidity contract for security vulnerabilities",
                "parameters": [
                    {
                        "description": "Contract source code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/analysis.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/static-analysis/{sourceHash}": {
            "get": {
                "description": "Retrieves static analysis result from cache or database using source hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "static-analysis"
                ],
                "summary": "Get static analysis result by source hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source hash",
                        "name": "sourceHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/test-cases/generate": {
            "post": {
                "description": "Generates test cases for Solidity contract using Gemini AI with comprehensive analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "testcase-generation"
                ],
                "summary": "Generate test cases for Solidity contract",
                "parameters": [
                    {
                        "description": "Contract source code and test preferences",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/testcase_generation.TestCaseGenerateRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/test-cases/{sourceHash}": {
            "get": {
                "description": "Gets test cases for Solidity contract by source hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "testcase-generation"
                ],
                "summary": "Get test cases for Solidity contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source hash",
                        "name": "sourceHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "analysis.AnalyzeRequest": {
            "type": "object",
            "required": [
                "source_code"
            ],
            "properties": {
                "source_code": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "testcase_generation.TestCaseGenerateRequest": {
            "type": "object",
            "required": [
                "source_code",
                "test_framework",
                "test_language"
            ],
            "properties": {
                "source_code": {
                    "type": "string",
                    "minLength": 1
                },
                "test_framework": {
                    "type": "string",
                    "example": "hardhat"
                },
                "test_language": {
                    "type": "string",
                    "example": "javascript"
                }
            }
        }
    }
}